drop custom project setup, add prisma config
