+++ backend/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 package-lock.json
 build
+.env
