+++ backend/package.json
@@ -28,8 +28,9 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/express": "^4.17.21",
     "@types/jest": "^28.1.8",
-    "@types/node": "^18.6.1",
+    "@types/node": "^18.19.33",
     "@typescript-eslint/eslint-plugin": "^5.31.0",
     "@typescript-eslint/parser": "^5.31.0",
     "eslint": "^8.20.0",
@@ -48,6 +49,12 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
-    "jest-cucumber": "^3.0.1"
+    "@prisma/client": "^5.14.0",
+    "@types/cors": "^2.8.17",
+    "cors": "^2.8.5",
+    "express": "^4.19.2",
+    "jest-cucumber": "^3.0.1",
+    "pg": "^8.11.5",
+    "prisma": "^5.14.0"
   }
 }
