--- backend/src/index.spec.ts
@@ -1,5 +1,0 @@
-
-describe('', () => {
-
-
-})
