+++ backend/src/seed.ts
@@ -1,0 +1,30 @@
+import { PrismaClient } from '@prisma/client'
+
+const prisma = new PrismaClient()
+
+async function main() {
+    await prisma.user.create({
+      data: {
+        username: 'Alice',
+        email: 'alice@prisma.io',
+      },
+    })
+  
+    const allUsers = await prisma.user.findMany({
+      include: {
+        posts: true,
+        profile: true,
+      },
+    })
+    console.dir(allUsers, { depth: null })
+  }
+
+  main()
+  .then(async () => {
+    await prisma.$disconnect()
+  })
+  .catch(async (e) => {
+    console.error(e)
+    await prisma.$disconnect()
+    process.exit(1)
+  })
