--- package.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": {
-    "@prisma/client": "^5.14.0",
-    "@types/cors": "^2.8.17",
-    "express": "^4.19.2",
-    "pg": "^8.11.5",
-    "prisma": "^5.14.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.21"
-  }
-}
